build: replace :stores:leveldb with :stores:rocksdb
Swaps the kotlin-leveldb engine for io.maryk.rocksdb:rocksdb-multiplatform
across the store module, version-catalog, sample, and README. Targets are
the same minus iosX64 / watchosX64 / tvosX64, which rocksdb-multiplatform
no longer publishes.

Drives three latent build-config bumps:

- Wrapper 8.14 -> 9.4.0. AGP-KMP-library 9.1.1 needs Gradle 9.3.1+; was
  working only because precompiled-script-plugin accessors were cached.
- compileSdk 35 -> 36 in kotlin-multiplatform-with-android-convention.
  rocksdb-android AAR refuses to link against 35.
- jvmToolchain 8 -> 11 in kotlin-multiplatform-convention. rocksdb-multiplatform
  JVM artifact is class-file v55 (Java 11+).

useLocalLevelDb includeBuild block and dead .gitmodules entry removed.
