+++ samples/ktor-server/src/main/kotlin/com/github/lamba92/kotlin/document/store/samples/ktor/server/Main.kt
@@ -3,7 +3,7 @@ package com.github.lamba92.kotlin.document.store.samples.ktor.server
 import com.github.lamba92.kotlin.document.store.core.KotlinDocumentStore
 import com.github.lamba92.kotlin.document.store.core.getObjectCollection
 import com.github.lamba92.kotlin.document.store.samples.User
-import com.github.lamba92.kotlin.document.store.stores.leveldb.LevelDBStore
+import com.github.lamba92.kotlin.document.store.stores.rocksdb.RocksDBStore
 import io.ktor.server.cio.CIO
 import io.ktor.server.engine.embeddedServer
 import kotlinx.coroutines.coroutineScope
@@ -11,7 +11,7 @@ import kotlinx.coroutines.coroutineScope
 suspend fun main() {
     val dbPath = System.getenv("DB_PATH") ?: error("DB_PATH environment variable not set")
     coroutineScope {
-        val db = KotlinDocumentStore(LevelDBStore.open(dbPath))
+        val db = KotlinDocumentStore(RocksDBStore.open(dbPath))
         val userCollection = db.getObjectCollection<User>("users")
 
         userCollection.createIndex("name")
