+++ settings.gradle.kts
@@ -1,11 +1,5 @@
 @file:Suppress("UnstableApiUsage")
 
-import java.nio.file.Path
-import kotlin.io.path.absolutePathString
-import kotlin.io.path.exists
-import kotlin.io.path.isDirectory
-
-
 plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "1.0.0"
     id("com.gradle.develocity") version "4.0.2"
@@ -28,54 +22,11 @@ include(
     ":samples:kmp-app",
     ":samples:ktor-server",
     ":stores:browser",
-    ":stores:leveldb",
+    ":stores:rocksdb",
     ":stores:mvstore",
     ":tests",
     ":version-catalog",
 )
-
-val levelDbPath: Path = file("../kotlin-leveldb").toPath()
-
-val localLeveldbExists = levelDbPath.isDirectory() && levelDbPath.resolve("settings.gradle.kts").exists()
-
-val useLocalLevelDb: Boolean? by settings
-
-if (localLeveldbExists && !isCi && useLocalLevelDb == true) {
-    includeBuild(levelDbPath.absolutePathString()) {
-        val endings = listOf(
-            "jvm",
-            "js",
-
-            "mingwx64",
-            "linuxx64",
-            "linuxarm64",
-            "macosx64",
-            "macosarm64",
-
-            "iosarm64",
-            "iosx64",
-            "iosSimulatorarm64",
-            "watchosarm64",
-            "watchosx64",
-            "watchosSimulatorarm64",
-            "tvosarm64",
-            "tvosx64",
-            "tvosSimulatorarm64",
-            "androidarm64",
-
-            "android",
-            "androidarm32",
-            "androidx86",
-            "androidx64"
-        )
-        dependencySubstitution {
-            substitute(module("com.github.lamba92:kotlin-leveldb")).using(project(":"))
-            endings.forEach {
-                substitute(module("com.github.lamba92:kotlin-leveldb-$it")).using(project(":"))
-            }
-        }
-    }
-}
 
 develocity {
     buildScan {
