--- stores/leveldb/src/androidInstrumentedTest/kotlin/com/github/lamba92/kotlin/document/store/tests/stores/leveldb/LeveldbTests.android.kt
@@ -1,12 +1,0 @@
-package com.github.lamba92.kotlin.document.store.tests.stores.leveldb
-
-import androidx.test.platform.app.InstrumentationRegistry
-
-actual val DB_PATH: String
-    get() =
-        InstrumentationRegistry
-            .getInstrumentation()
-            .targetContext
-            .filesDir
-            .resolve("testdb")
-            .absolutePath
