--- stores/leveldb/src/androidUnitTest/kotlin/com/github/lamba92/kotlin/document/store/tests/stores/leveldb/LeveldbTests.android.kt
@@ -1,4 +1,0 @@
-package com.github.lamba92.kotlin.document.store.tests.stores.leveldb
-
-actual val DB_PATH: String
-    get() = error("No unit tests are executed on Android")
