--- stores/leveldb/src/nativeDesktopTest/kotlin/com/github/lamba92/kotlin/document/store/tests/stores/leveldb/LeveldbTests.nativeDesktop.kt
@@ -1,6 +1,0 @@
-package com.github.lamba92.kotlin.document.store.tests.stores.leveldb
-
-import kotlinx.cinterop.toKString
-
-actual val DB_PATH: String
-    get() = platform.posix.getenv("DB_PATH")?.toKString() ?: error("DB_PATH not set")
