+++ stores/rocksdb/src/androidInstrumentedTest/kotlin/com/github/lamba92/kotlin/document/store/tests/stores/rocksdb/RocksdbTests.android.kt
@@ -1,0 +1,12 @@
+package com.github.lamba92.kotlin.document.store.tests.stores.rocksdb
+
+import androidx.test.platform.app.InstrumentationRegistry
+
+actual val DB_PATH: String
+    get() =
+        InstrumentationRegistry
+            .getInstrumentation()
+            .targetContext
+            .filesDir
+            .resolve("testdb")
+            .absolutePath
