+++ stores/rocksdb/src/androidNativeTest/kotlin/com/github/lamba92/kotlin/document/store/tests/stores/rocksdb/RocksdbTests.androidNative.kt
@@ -1,0 +1,6 @@
+package com.github.lamba92.kotlin.document.store.tests.stores.rocksdb
+
+actual val DB_PATH: String
+    get() = error()
+
+private fun error(): Nothing = error("Kotlin/Native has no tests suite for Android Native")
