+++ stores/rocksdb/src/androidUnitTest/kotlin/com/github/lamba92/kotlin/document/store/tests/stores/rocksdb/RocksdbTests.android.kt
@@ -1,0 +1,4 @@
+package com.github.lamba92.kotlin.document.store.tests.stores.rocksdb
+
+actual val DB_PATH: String
+    get() = error("No unit tests are executed on Android")
