+++ stores/rocksdb/src/jvmTest/kotlin/com/github/lamba92/kotlin/document/store/tests/stores/rocksdb/RocksdbTests.jvm.kt
@@ -1,0 +1,4 @@
+package com.github.lamba92.kotlin.document.store.tests.stores.rocksdb
+
+actual val DB_PATH: String
+    get() = System.getenv("DB_PATH") ?: error("DB_PATH not set")
