+++ stores/rocksdb/src/nativeDesktopTest/kotlin/com/github/lamba92/kotlin/document/store/tests/stores/rocksdb/RocksdbTests.nativeDesktop.kt
@@ -1,0 +1,6 @@
+package com.github.lamba92.kotlin.document.store.tests.stores.rocksdb
+
+import kotlinx.cinterop.toKString
+
+actual val DB_PATH: String
+    get() = platform.posix.getenv("DB_PATH")?.toKString() ?: error("DB_PATH not set")
